requirements.txt
python-telegram-bot
requests
epicstore_api